Update number_of_bins.m
The function was trying to find such a number of bins that N_runs would be divisible by it without remainder. In case not a round number, in an extreme case, it would return 1 bin. The corresponding while loop is deleted, the resulting significance indices for all test datasets are fine.